server/requirements.txt
docx==0.8.11
python-pptx==0.6.18
openpyxl==3.0.9
PyMuPDF==1.19.6
requests==2.26.0
pytesseract==0.3.8
Pillow==8.4.0
opencv-python==4.5.3.20210920
numpy==1.21.2